rules-engines/python/eval_engine/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.12.9
    # via semantic-kernel
aioice==0.10.1
    # via aiortc
aiortc==1.13.0
    # via semantic-kernel
aiosignal==1.3.2
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   openai
attrs==25.3.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
av==14.4.0
    # via aiortc
azure-ai-agents==1.0.0
    # via
    #   azure-ai-projects
    #   semantic-kernel
azure-ai-inference==1.0.0b9
    # via semantic-kernel
azure-ai-projects==1.0.0b11
    # via semantic-kernel
azure-common==1.1.28
    # via azure-search-documents
azure-core==1.34.0
    # via
    #   azure-ai-agents
    #   azure-ai-inference
    #   azure-ai-projects
    #   azure-core-tracing-opentelemetry
    #   azure-cosmos
    #   azure-identity
    #   azure-search-documents
    #   azure-storage-blob
azure-core-tracing-opentelemetry==1.0.0b12
    # via semantic-kernel
azure-cosmos==4.9.0
    # via
    #   -r requirements.in
    #   semantic-kernel
azure-identity==1.23.0
    # via
    #   -r requirements.in
    #   semantic-kernel
azure-search-documents==11.6.0b12
    # via semantic-kernel
azure-storage-blob==12.25.1
    # via azure-ai-projects
black==25.1.0
    # via -r requirements.in
certifi==2025.4.26
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via
    #   aiortc
    #   cryptography
    #   pylibsrtp
chardet==5.2.0
    # via prance
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via black
cloudevents==1.12.0
    # via semantic-kernel
coverage[toml]==7.8.2
    # via pytest-cov
cryptography==45.0.3
    # via
    #   aiortc
    #   azure-identity
    #   azure-storage-blob
    #   msal
    #   pyjwt
    #   pyopenssl
defusedxml==0.7.1
    # via semantic-kernel
deprecation==2.1.0
    # via cloudevents
distro==1.9.0
    # via openai
dnspython==2.7.0
    # via aioice
docopt==0.6.2
    # via -r requirements.in
duckdb==1.3.0
    # via -r requirements.in
faker==37.3.0
    # via -r requirements.in
frozenlist==1.6.2
    # via
    #   aiohttp
    #   aiosignal
google-crc32c==1.7.1
    # via aiortc
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   -r requirements.in
    #   openai
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
ifaddr==0.2.0
    # via aioice
importlib-metadata==8.7.0
    # via opentelemetry-api
iniconfig==2.1.0
    # via pytest
isodate==0.7.2
    # via
    #   azure-ai-agents
    #   azure-ai-inference
    #   azure-ai-projects
    #   azure-search-documents
    #   azure-storage-blob
    #   openapi-core
jinja2==3.1.6
    # via
    #   -r requirements.in
    #   semantic-kernel
jiter==0.10.0
    # via openai
jsonschema==4.24.0
    # via
    #   openapi-core
    #   openapi-schema-validator
    #   openapi-spec-validator
jsonschema-path==0.3.4
    # via
    #   openapi-core
    #   openapi-spec-validator
jsonschema-specifications==2025.4.1
    # via
    #   jsonschema
    #   openapi-schema-validator
lazy-object-proxy==1.11.0
    # via openapi-spec-validator
markupsafe==3.0.2
    # via
    #   jinja2
    #   werkzeug
more-itertools==10.7.0
    # via openapi-core
msal==1.32.3
    # via
    #   azure-identity
    #   msal-extensions
msal-extensions==1.3.1
    # via azure-identity
multidict==6.4.4
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.1.0
    # via black
nest-asyncio==1.6.0
    # via semantic-kernel
numpy==2.2.6
    # via
    #   scipy
    #   semantic-kernel
openai==1.84.0
    # via
    #   -r requirements.in
    #   semantic-kernel
openapi-core==0.19.5
    # via semantic-kernel
openapi-schema-validator==0.6.3
    # via
    #   openapi-core
    #   openapi-spec-validator
openapi-spec-validator==0.7.1
    # via openapi-core
opentelemetry-api==1.34.0
    # via
    #   azure-core-tracing-opentelemetry
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
    #   semantic-kernel
opentelemetry-sdk==1.34.0
    # via semantic-kernel
opentelemetry-semantic-conventions==0.55b0
    # via opentelemetry-sdk
packaging==25.0
    # via
    #   black
    #   deprecation
    #   prance
    #   pytest
parse==1.20.2
    # via openapi-core
pathable==0.4.4
    # via jsonschema-path
pathspec==0.12.1
    # via black
platformdirs==4.3.8
    # via black
pluggy==1.6.0
    # via pytest
prance==25.4.8.0
    # via semantic-kernel
propcache==0.3.1
    # via
    #   aiohttp
    #   yarl
protobuf==6.31.1
    # via semantic-kernel
psutil==7.0.0
    # via -r requirements.in
pybars4==0.9.13
    # via semantic-kernel
pycparser==2.22
    # via cffi
pydantic==2.11.5
    # via
    #   -r requirements.in
    #   openai
    #   pydantic-settings
    #   semantic-kernel
pydantic-core==2.33.2
    # via pydantic
pydantic-settings==2.9.1
    # via semantic-kernel
pyee==13.0.0
    # via aiortc
pygments==2.19.1
    # via pytest
pyjwt[crypto]==2.10.1
    # via
    #   msal
    #   pyjwt
pylibsrtp==0.12.0
    # via aiortc
pymeta3==0.5.1
    # via pybars4
pyopenssl==25.1.0
    # via aiortc
pytest==8.4.0
    # via
    #   -r requirements.in
    #   pytest-asyncio
    #   pytest-cov
pytest-asyncio==1.0.0
    # via -r requirements.in
pytest-cov==6.1.1
    # via -r requirements.in
python-dotenv==1.1.0
    # via
    #   -r requirements.in
    #   pydantic-settings
pyyaml==6.0.2
    # via jsonschema-path
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-path
    #   jsonschema-specifications
requests==2.32.3
    # via
    #   azure-core
    #   jsonschema-path
    #   msal
    #   prance
rfc3339-validator==0.1.4
    # via openapi-schema-validator
rpds-py==0.25.1
    # via
    #   jsonschema
    #   referencing
ruamel-yaml==0.18.13
    # via prance
ruamel-yaml-clib==0.2.12
    # via ruamel-yaml
scipy==1.15.3
    # via semantic-kernel
semantic-kernel[azure]==1.32.2
    # via -r requirements.in
six==1.17.0
    # via
    #   azure-core
    #   rfc3339-validator
sniffio==1.3.1
    # via
    #   anyio
    #   openai
tqdm==4.67.1
    # via openai
typing-extensions==4.14.0
    # via
    #   azure-ai-agents
    #   azure-ai-inference
    #   azure-ai-projects
    #   azure-core
    #   azure-cosmos
    #   azure-identity
    #   azure-search-documents
    #   azure-storage-blob
    #   openai
    #   openapi-core
    #   opentelemetry-api
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
    #   pydantic
    #   pydantic-core
    #   pyee
    #   semantic-kernel
    #   typing-inspection
typing-inspection==0.4.1
    # via
    #   pydantic
    #   pydantic-settings
tzdata==2025.2
    # via faker
urllib3==2.4.0
    # via requests
websockets==15.0.1
    # via semantic-kernel
werkzeug==3.1.1
    # via openapi-core
yarl==1.20.0
    # via aiohttp
zipp==3.22.0
    # via importlib-metadata
